pyproject.toml
[project]
name = "phos"
version = "0.1.0"
description = ""
authors = [
    {name = "OstarkovSN",email = "ostarkovstepan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "plotly (>=6.1.1,<7.0.0)",
    "kaleido (==0.2.1)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyinstaller (>=6.13.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
